Clase 4
Se crea una nueva calase ArrayList
